requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
email_validator
SQLAlchemy
WTForms
python-dotenv
openai
